Remove manual installation steps
The manual installation instructions are out of date and have diverged significantly enough that we dont want to keep updating them. The install script generator is the recommended way of passing these arguments to the msis so we only include instructions for this method. If users want to modify any parameters, they can modify the `install.bat`

[skip ci]